Invoke-Command $using improvements
Always use the v5.1 $using: logic when dealing with any connection type
that is not based on the WSManConnectionInfo. This allows the caller to
use the more advanced $using features like vars with special names,
$using vars inside sub scopes and more. WSMan based connections might
still communicate with older versions but as the oldest supported
version is v3 on Server 2012 we don't need to keep the default as v2 to
support some of the newer features.